feat: implement  role-based access control, admin page, schema validation updates, and extend form component capabilities

+++ server/api/admin/account/[id].delete.ts
@@ -1,6 +1,24 @@
 import { getService } from '#bs/services/core/account';
+import { checkRoutePermissions } from '#bs/utils/check_route_permissions';
+
+defineRouteMeta({
+  openAPI: {
+    tags: ['Base Admin'],
+    description: 'Delete account by ID.',
+    responses: {
+      200: {
+        description: 'Success response'
+      },
+      403: {
+        description: 'Not authorized'
+      }
+    }
+  }
+})
 
 export default defineEventHandler(async (event) => {
+  await checkRoutePermissions(event, ['account.crud.delete']);
+
   const id = getRouterParam(event, 'id') || '';
   const service = getService();
   await service.delete(id);
